+++ common/src/main/java/com/saunhardy/afkstatus/command/AFKCommand.java
@@ -96,7 +96,7 @@ public class AFKCommand {
 
                                 if (Config.systemMessages()) {
                                     String msg = player.getName().getString() + " is now AFK.";
-                                    var server = player.getServer();
+                                    var server = player.level().getServer();
                                     if (server != null) {
                                         server.getPlayerList().broadcastSystemMessage(
                                                 Component.literal(msg).withStyle(style -> style.withColor(getConfiguredColor())),