+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## Version 1.1.0
 
 ### Added
-- Added Minecraft 1.21.8 support; Forge, NeoForge, and Fabric jars are now available
+- Added Minecraft 1.21.9 support; Forge, NeoForge, and Fabric jars are now available